hs - update tests
